Created a sample enemy with the enemy movement script
The enemy which i added to the main scene for reference, includes a script called Enemy_Movement. What it does is there is a left limit which can be set a value and it is how far it will travel left of it's starting position, and likewise with the right. The speed does the obvious: faster/slower movement.